+++ R/block.R
@@ -73,10 +73,9 @@ print.roxy_block <- function(x, ...) {
   cat_line("  ", obj[-1])
 }
 
-block_create <- function(tokens, call, srcref,
-                         global_options = list()) {
+block_create <- function(tokens, call, srcref) {
 
-  tags <- parse_tags(tokens, global_options = global_options)
+  tags <- parse_tags(tokens)
   if (length(tags) == 0) return()
 
   roxy_block(tags,
@@ -86,18 +85,13 @@ block_create <- function(tokens, call, srcref,
   )
 }
 
-block_set_env <- function(block, env,
-                          global_options = list()
-                          ) {
-
-  block <- block_evaluate(block, env, global_options = global_options)
+block_set_env <- function(block, env) {
+  block <- block_evaluate(block, env)
   block <- block_find_object(block, env)
   block
 }
 
-block_evaluate <- function(block, env,
-                           global_options = list()
-                           ) {
+block_evaluate <- function(block, env) {
 
   tags <- block_get_tags(block, "eval")
   if (length(tags) == 0) {
@@ -119,7 +113,7 @@ block_evaluate <- function(block, env,
     file = block$file,
     offset = block$line
   )
-  tags <- lapply(tokens, parse_tags, global_options = global_options)
+  tags <- lapply(tokens, parse_tags)
 
   # Interpolate results back into original locations
   block_replace_tags(block, "eval", tags)
@@ -199,8 +193,8 @@ block_replace_tags <- function(block, tags, values) {
 
 # parsing -----------------------------------------------------------------
 
-parse_tags <- function(tokens, global_options = list()) {
-  markdown_activate(tokens, global_options = global_options)
+parse_tags <- function(tokens) {
+  markdown_activate(tokens)
 
   tokens <- parse_description(tokens)
   compact(lapply(tokens, roxy_tag_parse))