+++ R/markdown-state.R
@@ -4,9 +4,6 @@ roxy_tag_parse.roxy_tag_md <- function(x) tag_toggle(x)
 #' @export
 roxy_tag_parse.roxy_tag_noMd <- function(x) tag_toggle(x)
 
-## If not specified in DESCRIPTION
-markdown_global_default <- FALSE
-
 ## Hacky global switch - this uses the fact that blocks are parsed
 ## one after the another, and that we set markdown on/off before each
 ## block
@@ -19,10 +16,8 @@ markdown_on <- function(value = NULL) {
   return(isTRUE(markdown_env$`markdown-support`))
 }
 
-markdown_activate <- function(tags, global_options = list()) {
+markdown_activate <- function(tags) {
   ## markdown on/off based on global flag and presence of @md & @nomd
-  ## we need to use markdown_global_default as well, because global_options
-  ## can be NULL, e.g. if called from parse_text()
 
   names <- purrr::map_chr(tags, "tag")
   has_md <- "md" %in% names
@@ -32,7 +27,7 @@ markdown_activate <- function(tags, global_options = list()) {
     roxy_tag_warning(tags[[1]], "Both @md and @noMd, no markdown parsing")
   }
 
-  md <- global_options$markdown %||% markdown_global_default
+  md <- roxy_meta_get("markdown", FALSE)
   if (has_md) md <- TRUE
   if (has_nomd) md <- FALSE