+++ R/object-usage.R
@@ -4,11 +4,11 @@ roxy_tag_parse.roxy_tag_usage <- function(x) {
 }
 
 # Prefer explicit \code{@@usage} to a \code{@@formals} list.
-topic_add_usage <- function(topic, block, old_usage = FALSE) {
+topic_add_usage <- function(topic, block) {
   tag <- block_get_tag(block, "usage")
 
   if (is.null(tag)) {
-    usage <- object_usage(block$object, old_usage = old_usage)
+    usage <- object_usage(block$object)
   } else if (tag$val == "NULL") {
     usage <- NULL
   } else {
@@ -26,42 +26,42 @@ format.rd_section_usage <- function(x, ...) {
 
 # object_usage ------------------------------------------------------------
 
-object_usage <- function(x, old_usage = FALSE) {
+object_usage <- function(x) {
   UseMethod("object_usage")
 }
 
-object_usage.default <- function(x, old_usage = FALSE) {
+object_usage.default <- function(x) {
   NULL
 }
 
-object_usage.data <- function(x, old_usage = FALSE) {
+object_usage.data <- function(x) {
   rd(x$alias)
 }
 
-object_usage.function <- function(x, old_usage = FALSE) {
-  function_usage(x$alias, formals(x$value), identity, old_usage = old_usage)
+object_usage.function <- function(x) {
+  function_usage(x$alias, formals(x$value), identity)
 }
 
 object_usage.s3generic <- object_usage.function
 
-object_usage.s3method <- function(x, old_usage = FALSE) {
+object_usage.s3method <- function(x) {
   method <- attr(x$value, "s3method")
   s3method <- function(name) {
     build_rd("\\method{", name, "}{", auto_backtick(method[2]), "}")
   }
-  function_usage(method[1], formals(x$value), s3method, old_usage = old_usage)
+  function_usage(method[1], formals(x$value), s3method)
 }
 
-object_usage.s4generic <- function(x, old_usage = FALSE) {
-  function_usage(x$value@generic, formals(x$value), identity, old_usage = old_usage)
+object_usage.s4generic <- function(x) {
+  function_usage(x$value@generic, formals(x$value), identity)
 }
 
-object_usage.s4method <- function(x, old_usage = FALSE) {
+object_usage.s4method <- function(x) {
   s4method <- function(name) {
     classes <- auto_backtick(as.character(x$value@defined))
     build_rd("\\S4method{", name, "}{", paste0(classes, collapse = ","), "}")
   }
-  function_usage(x$value@generic, formals(x$value), s4method, old_usage = old_usage)
+  function_usage(x$value@generic, formals(x$value), s4method)
 }
 
 # Function usage ----------------------------------------------------------
@@ -70,18 +70,18 @@ object_usage.s4method <- function(x, old_usage = FALSE) {
 # replacement, infix, regular
 # function, s3 method, s4 method, data
 
-function_usage <- function(name, formals, format_name = identity, old_usage = FALSE) {
+function_usage <- function(name, formals, format_name = identity) {
   if (is_replacement_fun(name) && !is_infix_fun(name)) {
     name <- str_replace(name, fixed("<-"), "")
     formals$value <- NULL
 
-    wrap_usage(name, format_name, formals, suffix = " <- value", old_usage = old_usage)
+    wrap_usage(name, format_name, formals, suffix = " <- value")
   } else if (is_infix_fun(name) && identical(format_name, identity)) {
     # If infix, and regular function, munge format
     arg_names <- names(formals)
     build_rd(arg_names[1], " ", format_name(name), " ", arg_names[2])
   } else {
-    wrap_usage(name, format_name, formals, old_usage = old_usage)
+    wrap_usage(name, format_name, formals)
   }
 }
 
@@ -122,7 +122,7 @@ args_call <- function(call, args) {
 #' @param formals List of function formals
 #' @param suffix Optional suffix, used for replacement functions
 #' @noRd
-wrap_usage <- function(name, format_name, formals, suffix = NULL, width = 80L, old_usage = FALSE) {
+wrap_usage <- function(name, format_name, formals, suffix = NULL, width = 80L) {
   # Quote non-syntactic names if no special formatting
   if (identical(format_name, identity)) {
     name <- auto_quote(name)
@@ -134,7 +134,7 @@ wrap_usage <- function(name, format_name, formals, suffix = NULL, width = 80L, o
   bare <- args_call(name, args)
   if (nchar(bare, type = "width") < width) {
     out <- args_call(format_name(name), args)
-  } else if (old_usage) {
+  } else if (roxy_meta_get("old_usage", FALSE)) {
     x <- args_call(format_name(name), args)
     out <- wrapUsage(x, width = as.integer(width))
   } else {
@@ -152,7 +152,7 @@ wrap_usage <- function(name, format_name, formals, suffix = NULL, width = 80L, o
 # helpers -----------------------------------------------------------------
 
 # used for testing
-call_to_usage <- function(code, old_usage = FALSE, env = pkg_env()) {
+call_to_usage <- function(code, env = pkg_env()) {
   obj <- call_to_object(!!enexpr(code), env)
-  gsub("\u{A0}", " ", as.character(object_usage(obj, old_usage = old_usage)))
+  gsub("\u{A0}", " ", as.character(object_usage(obj)))
 }