+++ R/tokenize.R
@@ -1,8 +1,5 @@
 # Returns list of roxy_blocks
-tokenize_file <- function(file,
-                          global_options = list(),
-                          srcref_path = NULL
-                          ) {
+tokenize_file <- function(file, srcref_path = NULL) {
   lines <- read_lines(file)
 
   parsed <- parse(
@@ -26,8 +23,7 @@ tokenize_file <- function(file,
       srcref = refs[has_tokens],
       tokens = tokens[has_tokens]
     ),
-    block_create,
-    global_options = global_options
+    block_create
   )
   purrr::compact(blocks)
 }